# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dstdelivery",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/add/toCart": {
            "post": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Add To Cart",
                "description": "",
                "operationId": "b4d9ac5db4bb0b0971d508b77b9542ca",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Please write product ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "totalQty",
                        "in": "query",
                        "description": "Please write totalQty",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/get/cart/products": {
            "get": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Get Cart Products",
                "description": "",
                "operationId": "5109a3eebfcb7725835c2243d3c3910f",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/delete/cart/products": {
            "delete": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Delete Cart Products",
                "description": "",
                "operationId": "e9a536f744f79a65e07d3584d3a875db",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Please write product ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/add/quantity/forOne/Product": {
            "post": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Add Quantity For One Product",
                "description": "",
                "operationId": "58a228bce11ba84a2c89a9aa6767c3e7",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Please write product ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "totalQty",
                        "in": "query",
                        "description": "Please Write quantity How much do you want to add",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}